import-duplicity: replace preexec_fn 3>&1 1>&2 with --log-file
Particularly since we already have a temp dir, just write the --log-fd
output to a file via --log-file instead. This avoids the need for the
preexec_fn redirections or the custom exo. It also revealed a
test-duplicity-import error on FreeBSD in CI that was not visible with
the previous arrangement.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>